requirements.txt
# Core
setuptools>=68.0.0

# Django
Django==6.0.2

# PostgreSQL adapter
psycopg2-binary==2.9.9

# Database URL parser for Render
dj-database-url==2.1.0

# Django REST Framework
djangorestframework==3.14.0

# JWT Authentication
djangorestframework-simplejwt==5.3.1

# CORS headers for React frontend
django-cors-headers==4.3.1

# Environment variables
python-decouple==3.8

# SendGrid for email sending
sendgrid==6.11.0

# Cloudinary for image uploads
cloudinary==1.36.0

# Email backend (for development)
django-environ==0.11.2

# For password hashing
argon2-cffi==23.1.0

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0

# Code quality
flake8==7.0.0
black==24.1.1
isort==5.13.2

# Timezone support
pytz==2024.1

# Production server
gunicorn==21.2.0

# Static files in production
whitenoise==6.6.0
